requirements.txt
pillow
fastapi
uvicorn
